Move interop host functions to separate mods.
