+++ utils/ayaka-runtime/src/plugin.rs
@@ -1,12 +1,15 @@
 //! The plugin utilities.
 
+mod fs_interop;
+mod log_interop;
+mod plugin_interop;
+
 use crate::*;
-use anyhow::{bail, Result};
+use anyhow::Result;
 use ayaka_plugin::*;
 use std::{
-    collections::{BTreeMap, BTreeSet, HashMap},
-    io::{Read, Seek, SeekFrom},
-    sync::{Arc, Mutex, RwLock, Weak},
+    collections::HashMap,
+    sync::{Arc, RwLock, Weak},
 };
 use stream_future::stream;
 use trylog::macros::*;
@@ -59,44 +62,6 @@ impl<M: RawModule> Module<M> {
     }
 }
 
-#[derive(Default)]
-struct FDMap {
-    map: BTreeMap<u64, Box<dyn SeekAndRead>>,
-    remain: BTreeSet<u64>,
-}
-
-impl FDMap {
-    pub fn open(&mut self, file: Box<dyn SeekAndRead>) -> u64 {
-        let new_fd = match self.remain.pop_first() {
-            Some(fd) => fd,
-            None => match self.map.last_key_value() {
-                Some((fd, _)) => fd + 1,
-                None => 1,
-            },
-        };
-        self.map.insert(new_fd, file);
-        new_fd
-    }
-
-    pub fn close(&mut self, fd: u64) {
-        self.map.remove(&fd);
-        self.remain.insert(fd);
-    }
-
-    pub fn read(&mut self, fd: u64, buf: &mut [u8]) -> std::io::Result<usize> {
-        self.map.get_mut(&fd).unwrap().read(buf)
-    }
-
-    pub fn seek(&mut self, fd: u64, pos: SeekFrom) -> std::io::Result<u64> {
-        self.map.get_mut(&fd).unwrap().seek(pos)
-    }
-}
-
-#[allow(unsafe_code)]
-unsafe impl Send for FDMap {}
-#[allow(unsafe_code)]
-unsafe impl Sync for FDMap {}
-
 /// The plugin runtime.
 pub struct Runtime<M: RawModule + Send + Sync + 'static = BackendModule> {
     modules: HashMap<String, Module<M>>,
@@ -116,121 +81,6 @@ pub enum LoadStatus {
 }
 
 impl<M: RawModule + Send + Sync + 'static> Runtime<M> {
-    fn new_linker(root_path: &VfsPath, handle: Arc<RwLock<Weak<Self>>>) -> Result<M::Linker> {
-        let mut store = M::Linker::new()?;
-
-        // log
-        let log_func = store.wrap(|(data,): (Record,)| {
-            let target = format!("{}::<plugin>::{}", module_path!(), data.target);
-            log::logger().log(
-                &log::Record::builder()
-                    .level(data.level)
-                    .target(&target)
-                    .args(format_args!("{}", data.msg))
-                    .module_path(data.module_path.as_deref())
-                    .file(data.file.as_deref())
-                    .line(data.line)
-                    .build(),
-            );
-            Ok(())
-        });
-        let log_flush_func = store.wrap(|_: ()| {
-            log::logger().flush();
-            Ok(())
-        });
-        store.import(
-            "log",
-            HashMap::from([
-                ("__log".to_string(), log_func),
-                ("__log_flush".to_string(), log_flush_func),
-            ]),
-        )?;
-
-        // plugin
-        let h = handle.clone();
-        let modules_func = store.wrap(move |_: ()| {
-            if let Some(this) = h.read().unwrap().upgrade() {
-                Ok(this.modules.keys().cloned().collect::<Vec<_>>())
-            } else {
-                bail!("Runtime hasn't been initialized.")
-            }
-        });
-        let h = handle;
-        let call_func = store.wrap_with(
-            move |mut handle, (module, name, args): (String, String, Vec<u8>)| {
-                if let Some(this) = h.read().unwrap().upgrade() {
-                    Ok(handle.call(
-                        this.modules[&module].module.inner(),
-                        &name,
-                        &args,
-                        |slice| Ok(slice.to_vec()),
-                    )?)
-                } else {
-                    bail!("Runtime hasn't been initialized.")
-                }
-            },
-        );
-        store.import(
-            "plugin",
-            HashMap::from([
-                ("__modules".to_string(), modules_func),
-                ("__call".to_string(), call_func),
-            ]),
-        )?;
-
-        // fs
-        let p = root_path.clone();
-        let read_dir_func = store.wrap(move |(path,): (String,)| {
-            Ok(p.join(&path[1..])?
-                .read_dir()
-                .map(|iter| iter.map(|p| p.as_str().to_string()).collect::<Vec<_>>())
-                .ok())
-        });
-        let p = root_path.clone();
-        let metadata_func = store.wrap(move |(path,): (String,)| {
-            Ok(p.join(&path[1..])?.metadata().map(FileMetadata::from).ok())
-        });
-        let p = root_path.clone();
-        let exists_func = store.wrap(move |(path,): (String,)| Ok(p.join(&path[1..])?.exists()?));
-
-        let fd_map = Arc::new(Mutex::new(FDMap::default()));
-        let p = root_path.clone();
-        let map = fd_map.clone();
-        let open_file_func = store.wrap(move |(path,): (String,)| {
-            let file = p.join(&path[1..])?.open_file();
-            Ok(file.map(|file| map.lock().unwrap().open(file)).ok())
-        });
-        let map = fd_map.clone();
-        let close_file_func = store.wrap(move |(fd,): (u64,)| {
-            map.lock().unwrap().close(fd);
-            Ok(())
-        });
-        let map = fd_map.clone();
-        let file_read_func = store.wrap_with(move |mut handle, (fd, ptr, len): (u64, i32, i32)| {
-            Ok(handle
-                .slice_mut(ptr, len, |buf| map.lock().unwrap().read(fd, buf))
-                .ok())
-        });
-        let map = fd_map;
-        let file_seek_func = store.wrap(move |(fd, pos): (u64, FileSeekFrom)| {
-            Ok(map.lock().unwrap().seek(fd, pos.into()).ok())
-        });
-        store.import(
-            "fs",
-            HashMap::from([
-                ("__read_dir".to_string(), read_dir_func),
-                ("__metadata".to_string(), metadata_func),
-                ("__exists".to_string(), exists_func),
-                ("__open_file".to_string(), open_file_func),
-                ("__close_file".to_string(), close_file_func),
-                ("__file_read".to_string(), file_read_func),
-                ("__file_seek".to_string(), file_seek_func),
-            ]),
-        )?;
-
-        Ok(store)
-    }
-
     /// Load plugins from specific directory and plugin names.
     ///
     /// The actual load folder will be `rel_to.join(dir)`.
@@ -277,7 +127,10 @@ impl<M: RawModule + Send + Sync + 'static> Runtime<M> {
 
         yield LoadStatus::CreateEngine;
         let handle = Arc::new(RwLock::new(Weak::new()));
-        let store = Self::new_linker(root_path, handle.clone())?;
+        let mut store = M::Linker::new()?;
+        log_interop::register(&mut store)?;
+        plugin_interop::register(&mut store, handle.clone())?;
+        fs_interop::register(&mut store, root_path)?;
         let mut runtime = Self::new();
 
         let total_len = paths.len();
